pyproject.toml
[project]
name = "mapgraph"
version = "0.2.0"
description = "一个强大的对象控制工具, 覆盖了配置管理、状态管理、上下文管理和生命周期管理。"
authors = [
    {name = "luxuncang", email = "luxuncang@qq.com"},
]
dependencies = [
    "python-dotenv>=1.0.1",
    "launart>=0.8.2",
    "typing-tool>=0.1.0",
]
requires-python = ">=3.10"
readme = "README.md"
license = {text = "MIT"}

[project.optional-dependencies]
aiohttp = [
    "aiohttp>=3.10.5",
]
httpx = [
    "httpx>=0.27.2",
]
uvicorn = [
    "uvicorn>=0.30.6",
]
starlette = [
    "starlette>=0.38.4",
]
tortoise = [
    "tortoise-orm>=0.21.6",
]
[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"


[tool.pdm]
distribution = true

[tool.pdm.dev-dependencies]
test = [
    "pytest>=8.3.1",
]

[tool.pdm.scripts]

test = "python -m pytest"